requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.2
sentence-transformers>=3.0.0
faiss-cpu==1.7.4
httpx==0.25.2
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
numpy>=1.26.0
